refactor: break out formatting functions
