+++ src/app/state/reducer.ts
@@ -5,6 +5,7 @@ import {
   getReverseCypher,
   type Cypher,
 } from "@/utils/cypher";
+import { combineQuote, normalizeQuote } from "@/utils/formatting";
 import { ActionType, type Action, type State } from "./types";
 
 export const initialState: State = {
@@ -125,24 +126,6 @@ function isCompleteWithError(state: State, newAnswer: Cypher) {
     applyCypher(state.encryptedQuote, newAnswer).length ===
       combineQuote(state.quote, state.author).length
   );
-}
-
-export function combineQuote(quote: string, author: string) {
-  return quote + " — " + author;
-}
-
-export function normalizeQuote(quote: string) {
-  return removeAccentsAndDiacritics(quote).toLocaleLowerCase("en-US");
-}
-
-function removeAccentsAndDiacritics(str: string) {
-  const str2 = str.replace(/ø/g, "o").replace(/Ø/g, "O");
-
-  if (str2.normalize) {
-    return str2.normalize("NFD").replace(/[\u0300-\u036f]/g, "");
-  }
-
-  return str2;
 }
 
 function createEmptyReverseCypher(cypher: Cypher) {
